[aon_timer] support 64-bit clock frequency
Opentitan earlgrey_1.0.0 added a second 4-byte register to the aon_timer. This
commit updates the driver API to allow setting the frequency to a 64-bit value.

+++ boards/opentitan/src/main.rs
@@ -96,7 +96,7 @@ impl EarlGreyConfig for ChipConfig {
     // Clock frequencies as of https://github.com/lowRISC/opentitan/pull/19479
     const CPU_FREQ: u32 = 24_000_000;
     const PERIPHERAL_FREQ: u32 = 6_000_000;
-    const AON_TIMER_FREQ: u32 = 250_000;
+    const AON_TIMER_FREQ: u64 = 250_000;
     const UART_BAUDRATE: u32 = 115200;
 }
 
@@ -107,7 +107,7 @@ impl EarlGreyConfig for ChipConfig {
     // Clock frequencies as of https://github.com/lowRISC/opentitan/pull/19479
     const CPU_FREQ: u32 = 24_000_000;
     const PERIPHERAL_FREQ: u32 = 6_000_000;
-    const AON_TIMER_FREQ: u32 = 250_000;
+    const AON_TIMER_FREQ: u64 = 250_000;
     const UART_BAUDRATE: u32 = 115200;
 }
 
@@ -116,7 +116,7 @@ impl EarlGreyConfig for ChipConfig {
     const NAME: &'static str = "silicon";
     const CPU_FREQ: u32 = 100_000_000;
     const PERIPHERAL_FREQ: u32 = 24_000_000;
-    const AON_TIMER_FREQ: u32 = 200_000;
+    const AON_TIMER_FREQ: u64 = 200_000;
     const UART_BAUDRATE: u32 = 115200;
 }
 
@@ -127,7 +127,7 @@ impl EarlGreyConfig for ChipConfig {
     // Clock frequencies as of https://github.com/lowRISC/opentitan/pull/19368
     const CPU_FREQ: u32 = 500_000;
     const PERIPHERAL_FREQ: u32 = 125_000;
-    const AON_TIMER_FREQ: u32 = 125_000;
+    const AON_TIMER_FREQ: u64 = 125_000;
     const UART_BAUDRATE: u32 = 7200;
 }
